match all filter refactor
